+++ inst/include/dqrng_distribution.h
@@ -32,14 +32,6 @@ using uniform_distribution = boost::random::uniform_real_distribution<double>;
 using normal_distribution = boost::random::normal_distribution<double>;
 using exponential_distribution = boost::random::exponential_distribution<double>;
 
-template<typename DIST, typename RES>
-inline RES generate(size_t n, rng64_t rng, const DIST& dist) {
-  RES result(n);
-  auto gen = std::bind(dist, std::ref(*rng));
-  std::generate(result.begin(), result.end(), gen);
-  return result;
-}
-
 inline double uniform01(uint64_t x) {
   // prefer high bits due to weakness of lowest bits for xoshiro/xoroshiro with used "+" scrambler
   return (x >> 11) * (1. / (UINT64_C(1) << 53));